+++ packages/core/src/graphql/graphiql.html.ts
